% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{allowed_entities}
\alias{allowed_entities}
\title{Allowed entities: Kolada metadata classes}
\usage{
allowed_entities()
}
\value{
A vector of names of allowed metadata entities, i.e. the correct
spelling of all allowed values of the \code{entity} parameter in \code{\link{get_metadata}}.
}
\description{
Allowed entities: Kolada metadata classes
}
