% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cast}
\alias{cast}
\title{Cast column types in data.table}
\usage{
cast(x, fun, cols)
}
\arguments{
\item{x}{data.table}

\item{fun}{type function (e.g. as.integer)}

\item{cols}{character vector with columns to convert}
}
\description{
Function to change a column type for data.table columns.
}
\examples{
\dontrun{
library(data.table)
dt <- data.table(mtcars)
cast(dt, as.character, c('vs', 'cyl'))
}
}

