% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_specs.R
\name{check_table_spec}
\alias{check_table_spec}
\title{Compare a Kolada metadata table to a minimal spec for a metadata entity.}
\usage{
check_table_spec(meta_df, entity)
}
\arguments{
\item{meta_df}{The metadata table to be compared.}

\item{entity}{The metadata entity expected.}
}
\description{
Compare a Kolada metadata table to a minimal spec for a metadata entity.
}
