% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{desc_glue_spec}
\alias{desc_glue_spec}
\title{Description specifications in {glue} format}
\usage{
desc_glue_spec(type)
}
\arguments{
\item{type}{The metadata entity to be described.}
}
\description{
Description specifications in {glue} format
}
