% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_kpi}
\alias{get_kpi}
\alias{get_kpi_groups}
\alias{get_ou}
\alias{get_municipality}
\alias{get_municipality_groups}
\title{Download metadata from the Kolada API}
\usage{
get_kpi(id = NULL, cache = FALSE, cache_location = tempdir)

get_kpi_groups(id = NULL, cache = FALSE, cache_location = tempdir)

get_ou(id = NULL, municipality = NULL, cache = FALSE, cache_location = tempdir)

get_municipality(id = NULL, cache = FALSE, cache_location = tempdir)

get_municipality_groups(id = NULL, cache = FALSE, cache_location = tempdir)
}
\arguments{
\item{id}{(Optional) One or several KPI IDs}

\item{cache}{Logical. If TRUE, downloaded data are stored to the local disk in the place specified by \code{cache_location}. If data is already present on the local disk, this data is returned instead of downloading data from the API.}

\item{cache_location}{Where to store and search for cached data. Can be a path to a directory or the name of any function that returns the path to a directory when called, like \code{link{getwd}}. Defaults to \code{\link{tempdir}}.}

\item{municipality}{(Optional) A string or vector of strings cantaining
municipality codes. If getting OU data, you can use this parameter to narrow
the search.}
}
\value{
Returns a tibble with metadata for the specified entity.
}
\description{
There are five different types of metadata entities in the Kolada database: "kpi", "kpi_groups", "municipality", "municipality_groups", and "ou". For every entity there is a corresponding function \code{get_ENTITY} which retrieves a table with the metadata for that entity. The \code{get_ENTITY} functions are thin wrappers around \code{\link{get_metadata}}.
}
\examples{

\dontrun{
# Download KPI table and store a cache copy of the results in
# your current working directory
kpi_df <- get_kpi(cache = "wd")

# Download KPI table and store a cache copy of the results in
# a tempfile (will expire when your R session ends)
kpi_df <- get_kpi(cache = TRUE)
}

}
