% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_values}
\alias{get_values}
\title{Get data from the Kolada API}
\usage{
get_values(
  kpi = NULL,
  municipality = NULL,
  period = NULL,
  ou = NULL,
  simplify = TRUE
)
}
\arguments{
\item{kpi}{What kpis should be fetched? Can be a single name or a vector of
names.}

\item{municipality}{For which municipalities should data be fetched? Can be a
single name or a vector of names.}

\item{period}{For what years should data be fetched? Can be one or more
four-digit integers or character strings.}

\item{ou}{(Optional) for what Operating Units should data be fetched? Only
available for certain KPIs.}

\item{simplify}{Whether to make results more human readable.}
}
\description{
Get data from the Kolada API.
}
