% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{kolada_cache_dir}
\alias{kolada_cache_dir}
\title{Get the rKolada cache directory}
\usage{
kolada_cache_dir()
}
\value{
A string containing the path to the cache directory.
}
\description{
Returns the path to the persistent cache directory used by rKolada.
The directory is created if it does not exist.
}
