% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.R
\name{kpi_bind_keywords}
\alias{kpi_bind_keywords}
\title{Add keyword columns to a Kolada KPI table}
\usage{
kpi_bind_keywords(kpi_df, n = 2)
}
\arguments{
\item{kpi_df}{A KPI table, e.g. as created by \code{\link{get_kpi}}.}

\item{n}{How many keyword columns should be added?}
}
\description{
Identify \code{n} keywords describing the KPI and add them as new columns. Keywords are inferred from the \code{title} field of the \code{kpi_df} argument.
}
\examples{
\dontrun{
kpi_df <- get_kpi() \%>\%
  kpi_add_keywords(n = 3)
}

}
