% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.R
\name{kpi_describe}
\alias{kpi_describe}
\title{Describe the KPIs in a Kolada KPI table}
\usage{
kpi_describe(
  kpi_df,
  max_n = 5,
  format = "inline",
  heading_level = 2,
  sub_heading_level = 3
)
}
\arguments{
\item{kpi_df}{A KPI table}

\item{max_n}{The maximum number of KPIs to describe.}

\item{format}{Output format. Can be one of "inline" or "md" (markdown).}

\item{heading_level}{The top heading level output format is "md".}

\item{sub_heading_level}{The sub heading level output format is "md".}
}
\description{
Describe the KPIs in a Kolada KPI table
}
