% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.R
\name{kpi_extract_ids}
\alias{kpi_extract_ids}
\title{Extract a vector of KPI ID strings from a Kolada KPI table}
\usage{
kpi_extract_ids(kpi_df)
}
\arguments{
\item{kpi_df}{A KPI table, e.g. as created by \code{\link{get_kpi}}.}
}
\description{
Extract a vector of KPI ID strings from a Kolada KPI table
}
