% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_groups.R
\name{kpi_grp_extract_ids}
\alias{kpi_grp_extract_ids}
\title{Extract KPI ID strings from a Kolada KPI group table}
\usage{
kpi_grp_extract_ids(kpi_grp_df)
}
\arguments{
\item{kpi_grp_df}{A Kolada KPI group table, as created by e.g. \code{get_kpi_groups}.}
}
\description{
Extract KPI ID strings from a Kolada KPI group table
}
