% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_groups.R
\name{kpi_grp_search}
\alias{kpi_grp_search}
\title{Search a Kolada KPI group table for group names}
\usage{
kpi_grp_search(kpi_grp_df, query)
}
\arguments{
\item{kpi_grp_df}{A Kolada KPI group table, as created by e.g. \code{get_kpi_groups}.}

\item{query}{A search term or a vector of search terms to filter by. Case insensitive.}
}
\description{
Search a Kolada KPI group table for group names
}
