% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_groups.R
\name{kpi_grp_unnest_kpis}
\alias{kpi_grp_unnest_kpis}
\title{Create a KPI table from a KPI group table}
\usage{
kpi_grp_unnest_kpis(kpi_grp_df)
}
\arguments{
\item{kpi_grp_df}{A Kolada KPI group table, as created by e.g. \code{get_kpi_groups}.}
}
\description{
Create a KPI table from a KPI group table
}
