% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{municipality_extract_ids}
\alias{municipality_extract_ids}
\title{Extract a vector of municipality ID strings from a Kolada municipality table}
\usage{
municipality_extract_ids(munic_df)
}
\arguments{
\item{munic_df}{A Kolada Municipality metadata table, as created by e.g. \code{get_municipality}.}
}
\description{
Extract a vector of municipality ID strings from a Kolada municipality table
}
