% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_grp.R
\name{municipality_grp_describe}
\alias{municipality_grp_describe}
\title{Describe the municipalitie in a Kolada Municipality Group metadata table}
\usage{
municipality_grp_describe(
  munic_grp_df,
  max_n = 5,
  format = "inline",
  heading_level = 2,
  sub_heading_level = heading_level + 1
)
}
\arguments{
\item{munic_grp_df}{A Kolada Municipality Group metadata table, as created by e.g. \code{get_municipality_groups}.}

\item{max_n}{The maximum number of KPI groups to describe.}

\item{format}{Output format. Can be one of "inline" or "md" (markdown).}

\item{heading_level}{The top heading level output format is "md".}

\item{sub_heading_level}{The sub heading level output format is "md".}
}
\description{
Print a human-readable description of each row of a Municipality Group metadata table, including member municipalities (up
to a maximum number of rows). Can be printed either directly to the R console
or used to populate a R markdown document, which can be useful for documentation
purposes.
}
\examples{
\dontrun{
munic_grp_df <- get_municipality_groups()

# Which municipality groups match the keyword "Arboga"?
munic_grp_df \%>\% municipality_grp_search("Arboga")
# 11 groups

# Which KPI groups contain the municipality "Arboga"?
munic_grp_df \%>\%
  municipality_grp_unnest() \%>\%
  municipality_search("arboga") \%>\%
  count(group_title, sort = TRUE)
# 92 groups
}

}
