% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_grp.R
\name{municipality_grp_search}
\alias{municipality_grp_search}
\title{Search a Kolada KPI group table for group names}
\usage{
municipality_grp_search(munic_grp_df, query)
}
\arguments{
\item{munic_grp_df}{A Kolada municipality group table, as created by e.g. \code{get_municipality_groups}.}

\item{query}{A search term or a vector of search terms to filter by. Case insensitive.}
}
\description{
Search a Kolada KPI group table for group names
}
