% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_grp.R
\name{municipality_grp_unnest_municipalities}
\alias{municipality_grp_unnest_municipalities}
\title{Create a municipality table from a Kolada municipality group table}
\usage{
municipality_grp_unnest_municipalities(munic_grp_df)
}
\arguments{
\item{munic_grp_df}{A Kolada municipality group table, as created by e.g. \code{get_municipality_groups}.}
}
\description{
Create a municipality table from a Kolada municipality group table
}
