% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{municipality_id_to_name}
\alias{municipality_id_to_name}
\title{Convert municipality ids to municipality names}
\usage{
municipality_id_to_name(munic_df, id, remove_na = FALSE)
}
\arguments{
\item{munic_df}{A Kolada Municipality metadata table, as created by e.g. \code{get_municipality}.}

\item{id}{ID ids of one or several municipalities. Allows repeats.}

\item{remove_na}{Should NA return values be removed?}
}
\description{
Convert municipality ids to municipality names
}
\examples{
munic_df <- get_municipality()
municipality_id_to_name(munic_df, c("1280", "1281", "0180", "1280"))

}
