% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{municipality_name_to_id}
\alias{municipality_name_to_id}
\title{Convert municipality names to municipality ids}
\usage{
municipality_name_to_id(munic_df, municipality, remove_na = FALSE)
}
\arguments{
\item{munic_df}{A Kolada Municipality metadata table, as created by e.g. \code{get_municipality}.}

\item{municipality}{Name of one or several municipalities. Case insensitive. Allows repeats.}

\item{remove_na}{Should NA return values be removed?}
}
\description{
Convert municipality names to municipality ids
}
\examples{
munic_df <- get_municipality()
municipality_name_to_id(munic_df, c("Malmö", "Lund", "Stockholm", "Malmö"))
munic_df \%>\% municipality_name_to_id(c("Malmö", "Lund", "Stockholm", "Malmö"))

}
