% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{municipality_search}
\alias{municipality_search}
\title{Search a Kolada municipality table}
\usage{
municipality_search(munic_df, query, column = NULL)
}
\arguments{
\item{munic_df}{A Kolada Municipality metadata table, as created by e.g. \code{get_municipality}.}

\item{query}{A search term or a vector of search terms to filter by. Case insensitive.}

\item{column}{(Optional) A string or character vector with the names of
columns in which to search for \code{query}.}
}
\description{
Search a Kolada municipality table
}
\examples{
\dontrun{
# Search for a single search term in a municipality table
munic_df <- get_municipality()
municipality_search(munic_df, "Malmö")

# Only keep columns with type == "K" (keep municipalities, drop regions)
munic_filter <- get_municipality(cache = TRUE) \%>\%
  municipality_search("K", column = "type")
}

}
