% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\name{ou_search}
\alias{ou_search}
\title{Search a Kolada Operating unit table}
\usage{
ou_search(ou_df, query, column = NULL)
}
\arguments{
\item{ou_df}{A Kolada Municipality metadata table, as created by e.g. \code{get_municipality}.}

\item{query}{A search term or a vector of search terms to filter by. Case insensitive.}

\item{column}{(Optional) A string or character vector with the names of
columns in which to search for \code{query}.}
}
\description{
Search a Kolada Operating unit table
}
\examples{
\dontrun{
# Search for a single search term in a municipality table
ou_df <- get_ou()
ou_search(ou_df, "skola")

# Only keep OU entities located in Stockholm municipality
ou_filter <- get_ou() \%>\%
  ou_search("Stockholm", column = "municipality")
}

}
