% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rKolada.R
\name{\%>\%}
\alias{\%>\%}
\title{re-export magrittr pipe operator}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
re-export magrittr pipe operator
}
