% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{values_minimize}
\alias{values_minimize}
\title{Simplify a Kolada value table}
\usage{
values_minimize(values_df)
}
\arguments{
\item{values_df}{A Kolada value table, as created by \code{\link{get_values}}.}
}
\description{
Simplify a Kolada value table
}
