% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{cache_handler}
\alias{cache_handler}
\title{Cache handler factory}
\usage{
cache_handler(entity, cache, cache_location)
}
\arguments{
\item{entity}{A keyword describing the data entity handled by the cache handler.}

\item{cache}{Whether data should be stored.}

\item{cache_location}{Where data should be stored (a path to a directory or a function that returns a path to a directory).}
}
\description{
Cache handler factory
}
