% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_order_metadata.R
\name{generate_kpi_metadata}
\alias{generate_kpi_metadata}
\title{Generate KPI metadata from a Kolada KPI object}
\usage{
generate_kpi_metadata(kpi_df)
}
\arguments{
\item{kpi_df}{A Kolada KPI tibble}
}
\value{
a Kolada KPI metadata table
}
\description{
Generate KPI metadata from a Kolada KPI object
}
