% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{kolada_available}
\alias{kolada_available}
\title{Check if the Kolada API is available}
\usage{
kolada_available()
}
\description{
Check if the Kolada API is available and if data from the API makes sense.
This is primarily useful for programming with rKolada and/or examples.
}
