% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_grp.R
\name{municipality_grp_describe}
\alias{municipality_grp_describe}
\title{Describe the KPIs in a Kolada KPI group table}
\usage{
municipality_grp_describe(
  munic_grp_df,
  max_n = 5,
  format = "inline",
  heading_level = 2,
  sub_heading_level = 3
)
}
\arguments{
\item{munic_grp_df}{A Kolada municipality group table, as created by e.g. \code{get_municipality_groups}.}

\item{max_n}{The maximum number of KPI groups to describe.}

\item{format}{Output format. Can be one of "inline" or "md" (markdown).}

\item{heading_level}{The top heading level output format is "md".}

\item{sub_heading_level}{The sub heading level output format is "md".}
}
\description{
Describe the KPIs in a Kolada KPI group table
}
