% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_grp.R
\name{municipality_grp_extract_ids}
\alias{municipality_grp_extract_ids}
\title{Extract municipality ID strings from a Kolada municipality group table}
\usage{
municipality_grp_extract_ids(munic_grp_df)
}
\arguments{
\item{munic_grp_df}{A Kolada municipality group table, as created by e.g.
\code{get_municipality_groups}.}
}
\value{
A vector of Municipality IDs.
}
\description{
This function is primarily intended as a convenient way to pass a (filtered)
Kolada municipality group metadata table to \code{\link{get_values}}. All IDs
of the municipalities contained in each group in the table are extracted.
}
