% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_grp.R
\name{municipality_grp_search}
\alias{municipality_grp_search}
\title{Search a Kolada Municipality Group metadata table for group names}
\usage{
municipality_grp_search(munic_grp_df, query)
}
\arguments{
\item{munic_grp_df}{A Kolada Municipality Group metadata table, as created by
e.g. \code{get_municipality_groups}.}

\item{query}{A search term or a vector of search terms to filter by. Case
insensitive.}
}
\value{
A Kolada Municipality Group metadata table
}
\description{
Search a Kolada Municipality Group metadata table. Only keep rows that
contain the search query. Searches group titles and group IDs. Note that this
function does not search for individual municipalities contained within
municipality groups! To search for KPIs within a KPI group, see examples
below for an example using \code{municipality_grp_unnest}.
}
