% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{values_legend}
\alias{values_legend}
\title{Create KPI long-form descriptions to add to a plot}
\usage{
values_legend(values_df, kpi_df)
}
\arguments{
\item{values_df}{A Kolada value table, as created by \code{\link{get_values}}.}

\item{kpi_df}{A KPI table, e.g. as created by \code{\link{get_kpi}}.}
}
\description{
Create KPI long-form descriptions to add to a plot
}
