% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{values_legend}
\alias{values_legend}
\title{Create KPI long-form descriptions to add to a plot}
\usage{
values_legend(values_df, kpi_df)
}
\arguments{
\item{values_df}{A Kolada value table, as created by
\code{\link{get_values}}.}

\item{kpi_df}{A KPI table, e.g. as created by \code{\link{get_kpi}}.}
}
\value{
A string which should be used as caption in a plot.
}
\description{
In a Kolada values table, only KPI ID names are preserved. But in plots you
often want to add a legend to explain what each KPI ID represents. But since
KPI explanations are mostly relatively wordy, ggplot2 legends are
under-dimensioned for this task. \code{values_legend} returns a string which
can conveniently be used as caption to a plot instead.
}
